+++ src/components/OurFriendsCards/OurFriendsCards.styled.jsx
@@ -54,21 +54,6 @@ const BoxContent = styled.div`
   text-align: start;
 `;
 
-const Title = styled.h2`
-  font-family: ${fonts.primary},
-  font-weight: 700;
-  margin-bottom: 40px;
-  font-size: 24px;
-  line-height: 1.38;
-   @media (min-width: ${screens.tablet}) {
-    font-size: 48px;
-    margin-bottom: 45px;
-   } ;
-   
-    @media (min-width: ${screens.desktop}) {
-      margin-bottom: 60px;
-   }`;
-
 const Name = styled.h3`
   font-family: ${fonts.primary};
   font-weight: 700;
@@ -80,6 +65,15 @@ const Name = styled.h3`
     font-size: 20px;
     line-height: 1.35;
   } ;
+`;
+
+const NameLink = styled.a`
+  font-family: inherit;
+  font-weight: inherit;
+  margin-bottom: inherit;
+  font-size: inherit;
+  line-height: inherit;
+  color: inherit;
 `;
 
 const WorkDaysList = styled.ul`
@@ -151,7 +145,6 @@ const WorkDaysTime = styled.span`
 
 const StyledComponents = {
   Card,
-  Title,
   List,
   BoxContent,
   Name,
@@ -163,6 +156,7 @@ const StyledComponents = {
   CardPic,
   WorkDaysItem,
   WorkDaysTime,
+  NameLink,
 };
 
 export default StyledComponents;
