+++ src/components/OurFriendsTitle/OurFriendsTitle.jsx
@@ -1,0 +1,8 @@
+import StyledComponents from './OurFriendsTitle.styled';
+const { Title } = StyledComponents;
+
+const OurFriendsTitle = () => {
+  return <Title>Our friend</Title>;
+};
+
+export default OurFriendsTitle;
